Added l515 sku
