+++ src/l500/CMakeLists.txt
@@ -2,8 +2,12 @@
 # Copyright(c) 2019 Intel Corporation. All Rights Reserved.
 target_sources(${LRS_TARGET}
     PRIVATE
-        "${CMAKE_CURRENT_LIST_DIR}/l500.cpp"
+        "${CMAKE_CURRENT_LIST_DIR}/l500-depth.cpp"
         "${CMAKE_CURRENT_LIST_DIR}/l500-private.cpp"
-        "${CMAKE_CURRENT_LIST_DIR}/l500.h"
+        "${CMAKE_CURRENT_LIST_DIR}/l500-color.cpp"
+        "${CMAKE_CURRENT_LIST_DIR}/l500-factory.cpp"
+        "${CMAKE_CURRENT_LIST_DIR}/l500-depth.h"
         "${CMAKE_CURRENT_LIST_DIR}/l500-private.h"
+        "${CMAKE_CURRENT_LIST_DIR}/l500-color.h"
+        "${CMAKE_CURRENT_LIST_DIR}/l500-factory.h"
 )
