+++ src/l500/l500-depth.cpp
@@ -5,7 +5,7 @@
 #include "device.h"
 #include "context.h"
 #include "stream.h"
-#include "l500.h"
+#include "l500-depth.h"
 #include "l500-private.h"
 #include "proc/decimation-filter.h"
 #include "proc/threshold.h" 
@@ -19,54 +19,35 @@
 
 namespace librealsense
 {
-    std::shared_ptr<device_interface> l500_info::create(std::shared_ptr<context> ctx,
-        bool register_device_notifications) const
+    using namespace ivcam2;
+
+    std::shared_ptr<uvc_sensor> l500_depth::create_depth_device(std::shared_ptr<context> ctx,
+        const std::vector<platform::uvc_device_info>& all_device_infos)
     {
-        return std::make_shared<l500_device>(ctx,
-                                             this->get_device_data(),
-                                             register_device_notifications);
+        auto&& backend = ctx->get_backend();
+
+        std::vector<std::shared_ptr<platform::uvc_device>> depth_devices;
+        for (auto&& info : filter_by_mi(all_device_infos, 0)) // Filter just mi=0, DEPTH
+            depth_devices.push_back(backend.create_uvc_device(info));
+
+        auto depth_ep = std::make_shared<l500_depth_sensor>(this, std::make_shared<platform::multi_pins_uvc_device>(depth_devices),
+            std::unique_ptr<frame_timestamp_reader>(new l500_timestamp_reader_from_metadata(backend.create_time_service())));
+
+        depth_ep->register_xu(depth_xu);
+        depth_ep->register_pixel_format(pf_z16_l500);
+        depth_ep->register_pixel_format(pf_confidence_l500);
+        depth_ep->register_pixel_format(pf_y8_l500);
+
+        depth_ep->register_option(RS2_OPTION_LASER_POWER,
+            std::make_shared<uvc_xu_option<int>>(
+                *depth_ep,
+                ivcam2::depth_xu,
+                ivcam2::IVCAM2_DEPTH_LASER_POWER, "Power of the l500 projector, with 0 meaning projector off"));
+
+        return depth_ep;
     }
 
-    std::vector<std::shared_ptr<device_info>> l500_info::pick_l500_devices(
-        std::shared_ptr<context> ctx,
-        std::vector<platform::uvc_device_info>& uvc,
-        std::vector<platform::usb_device_info>& usb)
-    {
-        std::vector<platform::uvc_device_info> chosen;
-        std::vector<std::shared_ptr<device_info>> results;
-
-        auto correct_pid = filter_by_product(uvc, { L500_PID, L500_PID2});
-        auto group_devices = group_devices_by_unique_id(correct_pid);
-        for (auto& group : group_devices)
-        {
-            if (!group.empty() && mi_present(group, 0))
-            {
-                auto depth = get_mi(group, 0);
-                platform::usb_device_info hwm;
-
-                if (ivcam2::try_fetch_usb_device(usb, depth, hwm))
-                {
-                    auto info = std::make_shared<l500_info>(ctx, group, hwm);
-                    chosen.push_back(depth);
-                    results.push_back(info);
-                }
-                else
-                {
-                    LOG_WARNING("try_fetch_usb_device(...) failed.");
-                }
-            }
-            else
-            {
-                LOG_WARNING("L500 group_devices is empty.");
-            }
-        }
-
-        trim_device_list(uvc, chosen);
-
-        return results;
-    }
-
-    std::vector<uint8_t> l500_device::get_raw_calibration_table() const
+    std::vector<uint8_t> l500_depth::get_raw_calibration_table() const
     {
         //WA untill fw will fix DPT_INTRINSICS_GET command
         command cmd_fx(0x01, 0xa00e0804, 0xa00e0808);
@@ -87,17 +68,16 @@ namespace librealsense
         return vec;
     }
 
-    l500_device::l500_device(std::shared_ptr<context> ctx,
-                             const platform::backend_device_group& group,
-                             bool register_device_notifications)
-        : device(ctx, group, register_device_notifications),
+    l500_depth::l500_depth(std::shared_ptr<context> ctx,
+                             const platform::backend_device_group& group)
+        :device(ctx, group),
         _depth_device_idx(add_sensor(create_depth_device(ctx, group.uvc_devices))),
         _depth_stream(new stream(RS2_STREAM_DEPTH)),
         _ir_stream(new stream(RS2_STREAM_INFRARED)),
         _confidence_stream(new stream(RS2_STREAM_CONFIDENCE))
     {
         _calib_table_raw = [this]() { return get_raw_calibration_table(); };
-        _zo_point = [&]() { return read_zo_point(); };
+        
         static const auto device_name = "Intel RealSense L500";
 
         using namespace ivcam2;
@@ -127,11 +107,6 @@ namespace librealsense
         register_info(RS2_CAMERA_INFO_PHYSICAL_PORT, group.uvc_devices.front().device_path);
         register_info(RS2_CAMERA_INFO_PRODUCT_ID, pid_hex_str);
 
-       
-        get_depth_sensor().register_option(RS2_OPTION_DEPTH_UNITS, std::make_shared<const_value_option>("Number of meters represented by a single depth unit",
-            lazy<float>([&]() {
-            return read_znorm(); })));
-
         get_depth_sensor().register_option(RS2_OPTION_LLD_TEMPERATURE,
             std::make_shared <l500_temperature_options>(_hw_monitor.get(), RS2_OPTION_LLD_TEMPERATURE));
 
@@ -140,7 +115,7 @@ namespace librealsense
 
         get_depth_sensor().register_option(RS2_OPTION_MA_TEMPERATURE,
             std::make_shared <l500_temperature_options>(_hw_monitor.get(), RS2_OPTION_MA_TEMPERATURE));
-        
+
         environment::get_instance().get_extrinsics_graph().register_same_extrinsics(*_depth_stream, *_ir_stream);
         environment::get_instance().get_extrinsics_graph().register_same_extrinsics(*_depth_stream, *_confidence_stream);
 
@@ -148,17 +123,17 @@ namespace librealsense
         register_stream_to_extrinsic_group(*_ir_stream, 0);
     }
 
-    void l500_device::create_snapshot(std::shared_ptr<debug_interface>& snapshot) const
+    void l500_depth::create_snapshot(std::shared_ptr<debug_interface>& snapshot) const
     {
         throw not_implemented_exception("create_snapshot(...) not implemented!");
     }
 
-    void l500_device::enable_recording(std::function<void(const debug_interface&)> record_action)
+    void l500_depth::enable_recording(std::function<void(const debug_interface&)> record_action)
     {
         throw not_implemented_exception("enable_recording(...) not implemented!");
     }
 
-    std::vector<tagged_profile> l500_device::get_profiles_tags() const
+    std::vector<tagged_profile> l500_depth::get_profiles_tags() const
     {
         std::vector<tagged_profile> tags;
 
@@ -169,43 +144,20 @@ namespace librealsense
         return tags;
     }
 
-    void l500_device::force_hardware_reset() const
+    void l500_depth::force_hardware_reset() const
     {
         command cmd(ivcam2::fw_cmd::HWReset);
         cmd.require_response = false;
         _hw_monitor->send(cmd);
     }
 
-    std::shared_ptr<matcher> l500_device::create_matcher(const frame_holder& frame) const
-    {
-        std::vector<std::shared_ptr<matcher>> depth_matchers;
-
-        std::vector<stream_interface*> streams = { _depth_stream.get(), _ir_stream.get(), _confidence_stream.get() };
-
-        for (auto& s : streams)
-        {
-            depth_matchers.push_back(std::make_shared<identity_matcher>(s->get_unique_id(), s->get_stream_type()));
-        }
-        std::vector<std::shared_ptr<matcher>> matchers;
-        if (!frame.frame->supports_frame_metadata(RS2_FRAME_METADATA_FRAME_COUNTER))
-        {
-            matchers.push_back(std::make_shared<timestamp_composite_matcher>(depth_matchers));
-        }
-        else
-        {
-            matchers.push_back(std::make_shared<frame_number_composite_matcher>(depth_matchers));
-        }
-
-        return std::make_shared<timestamp_composite_matcher>(matchers);
-    }
-
-    std::pair<int, int>l500_device::read_zo_point()
+    std::pair<int, int> l500_depth_sensor::read_zo_point()
     {
         if (auto ver = read_algo_version() >= 115)
         {
             const int zo_point_address = 0xa00e1b8c;
             command cmd(ivcam2::fw_cmd::MRD, zo_point_address, zo_point_address + 4);
-            auto res = _hw_monitor->send(cmd);
+            auto res = _owner->_hw_monitor->send(cmd);
             if (res.size() < 2)
             {
                 throw std::runtime_error("Invalid result size!");
@@ -216,11 +168,11 @@ namespace librealsense
         return { 0, 0 };
     }
 
-    int l500_device::read_algo_version()
+    int l500_depth_sensor::read_algo_version()
     {
         const int algo_version_address = 0xa0020bd8;
         command cmd(ivcam2::fw_cmd::MRD, algo_version_address, algo_version_address + 4);
-        auto res = _hw_monitor->send(cmd);
+        auto res = _owner->_hw_monitor->send(cmd);
         if (res.size() < 2)
         {
             throw std::runtime_error("Invalid result size!");
@@ -230,11 +182,11 @@ namespace librealsense
         return ver;
     }
 
-    float l500_device::read_baseline()
+    float l500_depth_sensor::read_baseline()
     {
         const int baseline_address = 0xa00e0868;
         command cmd(ivcam2::fw_cmd::MRD, baseline_address, baseline_address + 4);
-        auto res = _hw_monitor->send(cmd);
+        auto res = _owner->_hw_monitor->send(cmd);
         if (res.size() < 1)
         {
             throw std::runtime_error("Invalid result size!");
@@ -243,10 +195,10 @@ namespace librealsense
         return *data;
     }
 
-    float l500_device::read_znorm()
+    float l500_depth_sensor::read_znorm()
     {
         const int baseline_znorm = 0xa00e0b08;
-        auto res = _hw_monitor->send(command(ivcam2::fw_cmd::MRD, baseline_znorm, baseline_znorm + 4));
+        auto res = _owner->_hw_monitor->send(command(ivcam2::fw_cmd::MRD, baseline_znorm, baseline_znorm + 4));
         if (res.size() < 1)
         {
             throw std::runtime_error("Invalid result size!");
@@ -315,7 +267,7 @@ namespace librealsense
     {
         return _zo_point->second;
     }
-    processing_blocks l500_device::l500_depth_sensor::get_l500_recommended_proccesing_blocks(std::shared_ptr<option> zo_point_x, std::shared_ptr<option> zo_point_y)
+    processing_blocks l500_depth_sensor::get_l500_recommended_proccesing_blocks(std::shared_ptr<option> zo_point_x, std::shared_ptr<option> zo_point_y)
     {
         processing_blocks res;
         res.push_back(std::make_shared<zero_order>(zo_point_x, zo_point_y));
