+++ src/l500/l500-private.h
@@ -11,8 +11,6 @@ namespace librealsense
 {
     namespace ivcam2
     {
-        const uint16_t L500_PID = 0x0b0d;
-
         // L500 depth XU identifiers
         const uint8_t L500_HWMONITOR = 1;
 
@@ -58,6 +56,97 @@ namespace librealsense
         private:
             rs2_option _option;
             hw_monitor* _hw_monitor;
+        };
+
+        class l500_timestamp_reader : public frame_timestamp_reader
+        {
+            static const int pins = 3;
+            mutable std::vector<int64_t> counter;
+            std::shared_ptr<platform::time_service> _ts;
+            mutable std::recursive_mutex _mtx;
+        public:
+            l500_timestamp_reader(std::shared_ptr<platform::time_service> ts)
+                : counter(pins), _ts(ts)
+            {
+                reset();
+            }
+
+            void reset() override
+            {
+                std::lock_guard<std::recursive_mutex> lock(_mtx);
+                for (auto i = 0; i < pins; ++i)
+                {
+                    counter[i] = 0;
+                }
+            }
+
+            rs2_time_t get_frame_timestamp(const request_mapping& mode, const platform::frame_object& fo) override
+            {
+                std::lock_guard<std::recursive_mutex> lock(_mtx);
+                return _ts->get_time();
+            }
+
+            unsigned long long get_frame_counter(const request_mapping & mode, const platform::frame_object& fo) const override
+            {
+                std::lock_guard<std::recursive_mutex> lock(_mtx);
+                auto pin_index = 0;
+                if (mode.pf->fourcc == 0x5a313620) // Z16
+                    pin_index = 1;
+                else if (mode.pf->fourcc == 0x43202020) // Confidence
+                    pin_index = 2;
+
+                return ++counter[pin_index];
+            }
+
+            rs2_timestamp_domain get_frame_timestamp_domain(const request_mapping & mode, const platform::frame_object& fo) const override
+            {
+                return RS2_TIMESTAMP_DOMAIN_SYSTEM_TIME;
+            }
+        };
+
+        class l500_timestamp_reader_from_metadata : public frame_timestamp_reader
+        {
+            std::unique_ptr<l500_timestamp_reader> _backup_timestamp_reader;
+            bool one_time_note;
+            mutable std::recursive_mutex _mtx;
+            arithmetic_wraparound<uint32_t, uint64_t > ts_wrap;
+
+        protected:
+
+            bool has_metadata_ts(const platform::frame_object& fo) const
+            {
+                // Metadata support for a specific stream is immutable
+                const bool has_md_ts = [&] { std::lock_guard<std::recursive_mutex> lock(_mtx);
+                return ((fo.metadata != nullptr) && (fo.metadata_size >= platform::uvc_header_size) && ((byte*)fo.metadata)[0] >= platform::uvc_header_size);
+                }();
+
+                return has_md_ts;
+            }
+
+            bool has_metadata_fc(const platform::frame_object& fo) const
+            {
+                // Metadata support for a specific stream is immutable
+                const bool has_md_frame_counter = [&] { std::lock_guard<std::recursive_mutex> lock(_mtx);
+                return ((fo.metadata != nullptr) && (fo.metadata_size > platform::uvc_header_size) && ((byte*)fo.metadata)[0] > platform::uvc_header_size);
+                }();
+
+                return has_md_frame_counter;
+            }
+
+        public:
+            l500_timestamp_reader_from_metadata(std::shared_ptr<platform::time_service> ts) :_backup_timestamp_reader(nullptr), one_time_note(false)
+            {
+                _backup_timestamp_reader = std::unique_ptr<l500_timestamp_reader>(new l500_timestamp_reader(ts));
+                reset();
+            }
+
+            rs2_time_t get_frame_timestamp(const request_mapping& mode, const platform::frame_object& fo) override;
+
+            unsigned long long get_frame_counter(const request_mapping & mode, const platform::frame_object& fo) const override;
+
+            void reset() override;
+
+            rs2_timestamp_domain get_frame_timestamp_domain(const request_mapping & mode, const platform::frame_object& fo) const override;
         };
     } // librealsense::ivcam2
 } // namespace librealsense
